backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Processing
datasets==2.14.6
pandas==2.1.3
numpy==1.24.4

# LangChain Ecosystem
langchain-community==0.0.37
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langchain-openai==0.0.2
langchain-huggingface==0.0.2
langchain-qdrant==0.1.0
langchain-tavily==0.1.0
langgraph==0.0.66

# LangChain MCP Integration
langchain-mcp-adapters==0.1.0

# MCP Framework
mcp==1.0.0

# Vector Database
qdrant-client==1.7.0
sentence-transformers==2.2.2

# AI/ML Libraries
dspy-ai==2.4.9
groq==0.4.1

# Guardrails & Monitoring
portkey-ai==1.6.0

# Web Search
tavily-python==0.3.3

# API Dependencies
pydantic==2.5.0
typing-extensions==4.8.0

# Optional Dependencies for Enhanced Features
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2